feat: change alots

+++ auction/src/main/java/com/lcaohoanq/auction/repositories/AuctionStatusRepository.java
@@ -4,6 +4,6 @@ import com.lcaohoanq.auction.models.AuctionStatus;
 import com.lcaohoanq.auction.models.BidMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionStatusRepository extends JpaRepository<AuctionStatus, Long> {
+public interface AuctionStatusRepository extends JpaRepository<AuctionStatus, Integer> {
 
 }
