+++ auction/src/main/java/com/lcaohoanq/auction/repositories/BidMethodRepository.java
@@ -3,6 +3,6 @@ package com.lcaohoanq.auction.repositories;
 import com.lcaohoanq.auction.models.BidMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BidMethodRepository extends JpaRepository<BidMethod, Long> {
+public interface BidMethodRepository extends JpaRepository<BidMethod, Integer> {
 
 }
